Cloud/IOT/requirements.txt
numpy
kafka-python
requests
matplotlib
pillow